requirements.txt
neo4j
openai
langchain
langchain-core
langchain_openai
langchain-community
langsmith

python-dotenv




#2
langchain-experimental
yfiles-jupyter-graphs

#4
wikipedia